Notification refactoring
Modified EmailNotificationService, now it delegates to Event,Venue,
Service and Payment notification services.
Added EmailConstants class for email templates.
Added attendee reminder notification.
Added Organizer notification when provider accept/reject request.

Removed unnecessary interfaces to simplify service layer.
