+++ src/main/java/com/example/cdr/eventsmanagementsystem/Controller/ServiceController.java
@@ -6,7 +6,7 @@ import com.example.cdr.eventsmanagementsystem.DTO.Booking.Response.BookingDetail
 import com.example.cdr.eventsmanagementsystem.DTO.Service.ServicesDTO;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.BookingStatus;
-import com.example.cdr.eventsmanagementsystem.Service.Service.ServicesServiceInterface;
+import com.example.cdr.eventsmanagementsystem.Service.Service.ServicesService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -23,7 +23,7 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "Service", description = "Service provider APIs for managing services and bookings")
 public class
 ServiceController {
-    private final ServicesServiceInterface servicesService;
+    private final ServicesService servicesService;
 
     @Operation(summary = "Add a new service")
     @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
@@ -56,11 +56,5 @@ ServiceController {
     public Booking acceptOrRejectBooking(@PathVariable Long bookingId,@RequestParam BookingStatus status)   {
         return servicesService.respondToBookingRequests(bookingId, status);
     }
-//    @Operation(summary = "Cancel a booking as a service provider")
-//    @PreAuthorize("hasRole('" + RoleConstants.SERVICE_PROVIDER_ROLE + "')")
-//    @PostMapping(ServiceControllerConstants.CANCEL_SERVICE_BOOKING_URL)
-//    public ResponseEntity<Void> cancelBooking(@PathVariable Long bookingId)  {
-//        servicesService.cancelBooking(bookingId);
-//        return ResponseEntity.ok().build();
-//    }
+
 }
