+++ src/main/java/com/example/cdr/eventsmanagementsystem/EventsManagementSystemApplication.java
@@ -3,9 +3,11 @@ package com.example.cdr.eventsmanagementsystem;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableJpaAuditing(auditorAwareRef = "auditorAwareImpl")
+@EnableScheduling
 public class EventsManagementSystemApplication {
 
     public static void main(String[] args) {
