+++ src/main/java/com/example/cdr/eventsmanagementsystem/NotificationEvent/BookingUpdates/ServiceBookingUpdate.java
@@ -1,0 +1,5 @@
+package com.example.cdr.eventsmanagementsystem.NotificationEvent.BookingUpdates;
+
+import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
+
+public record ServiceBookingUpdate(Booking booking) {}
