--- src/main/java/com/example/cdr/eventsmanagementsystem/Service/Notifications/EmailNotificationChannel.java
@@ -1,23 +1,0 @@
-package com.example.cdr.eventsmanagementsystem.Service.Notifications;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class EmailNotificationChannel implements NotificationChannel {
-
-    private final JavaMailSender mailSender;
-
-    @Override
-    public void send(String to, String subject, String content) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setTo(to);
-        message.setSubject(subject);
-        message.setText(content);
-
-        mailSender.send(message);
-    }
-}
