+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/Notifications/Helper/NotificationHelper.java
@@ -3,11 +3,13 @@ package com.example.cdr.eventsmanagementsystem.Service.Notifications.Helper;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.BookingType;
 import com.example.cdr.eventsmanagementsystem.Model.User.BaseRoleEntity;
+import com.example.cdr.eventsmanagementsystem.Service.Auth.UserRoleHandler;
 import com.example.cdr.eventsmanagementsystem.Service.Auth.UserSyncService;
 import com.example.cdr.eventsmanagementsystem.Constants.NotificationConstants.EmailMessages;
-import com.example.cdr.eventsmanagementsystem.Service.Notifications.NotificationChannel;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Component;
 import java.util.Objects;
 
@@ -16,9 +18,9 @@ import java.util.Objects;
 @RequiredArgsConstructor
 public class NotificationHelper {
     private final UserSyncService userSyncService;
-    private final NotificationChannel emailNotification;
+    private final JavaMailSender mailSender;
 
-    private BaseRoleEntity getProvider(Booking booking, BookingType bookingType) {
+    public BaseRoleEntity getProvider(Booking booking, BookingType bookingType) {
         return switch (bookingType) {
             case VENUE -> booking.getVenue() != null ? booking.getVenue().getVenueProvider() : null;
             case SERVICE -> booking.getService() != null ? booking.getService().getServiceProvider() : null;
@@ -33,11 +35,14 @@ public class NotificationHelper {
         };
     }
     public BaseRoleEntity getUser(String userId, String context, Long bookingId) {
-        BaseRoleEntity user = userSyncService.findUserById(userId);
-        if (Objects.isNull(user)) {
-            log.error("User not found for {}: booking {}", context, bookingId);
+        for (UserRoleHandler<?> handler : userSyncService.getHandlers()) {
+            BaseRoleEntity user = handler.findUserById(userId);
+            if (user != null) {
+                return user;
+            }
         }
-        return user;
+        log.error("User not found for {}: booking {}", context, bookingId);
+        return null;
     }
     public void sendBookingConfirmation(Booking booking, String bookingTypeName) {
         BaseRoleEntity booker = getUser(booking.getBookerId(), "confirmation", booking.getId());
@@ -46,7 +51,7 @@ public class NotificationHelper {
         String content = String.format(EmailMessages.BOOKING_CONFIRMED,
                 booker.getFirstName(), bookingTypeName, booking.getId(), booking.getStripePaymentId());
 
-        emailNotification.send(booker.getEmail(),
+        send(booker.getEmail(),
                 "Booking Confirmed #" + booking.getId(), content);
 
     }
@@ -62,7 +67,7 @@ public class NotificationHelper {
                 provider.getFirstName(), resourceType, resourceName, booking.getId(),
                 booking.getStartTime(), booking.getEndTime(), booker.getFirstName());
 
-        emailNotification.send(provider.getEmail(),
+        send(provider.getEmail(),
                 "New Booking Request for Your " + resourceType.substring(0, 1).toUpperCase() + resourceType.substring(1),
                 content);
     }
@@ -79,9 +84,17 @@ public class NotificationHelper {
                 booking.getStartTime(), booking.getEndTime(), booker.getFullName(),
                 booker.getEmail(), reason);
 
-        emailNotification.send(provider.getEmail(),
+        send(provider.getEmail(),
                 "Booking Cancelled for Your " + resourceType.substring(0, 1).toUpperCase() + resourceType.substring(1),
                 content);
 
+    }
+    public void send(String to, String subject, String content) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(content);
+
+        mailSender.send(message);
     }
 }
