--- src/main/java/com/example/cdr/eventsmanagementsystem/Service/Notifications/NotificationChannel.java
@@ -1,5 +1,0 @@
-package com.example.cdr.eventsmanagementsystem.Service.Notifications;
-
-public interface NotificationChannel {
-    void send(String to, String subject, String content);
-}
