+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/Notifications/ServiceNotificationService.java
@@ -1,11 +1,17 @@
 package com.example.cdr.eventsmanagementsystem.Service.Notifications;
 
+import com.example.cdr.eventsmanagementsystem.Constants.NotificationConstants.EmailMessages;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.BookingType;
+import com.example.cdr.eventsmanagementsystem.Model.User.BaseRoleEntity;
 import com.example.cdr.eventsmanagementsystem.Service.Notifications.Helper.NotificationHelper;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
+
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ServiceNotificationService {
@@ -22,5 +28,20 @@ public class ServiceNotificationService {
     public void sendServiceCancellationEmail(Booking booking,String reason) {
         // Notify service provider
         notificationHelper.sendProviderCancellationNotification(booking, BookingType.SERVICE, "service", reason);
+    }
+    public void sendServiceBookingUpdateEmail(Booking booking) {
+        BaseRoleEntity organizer = notificationHelper.getUser(
+                booking.getBookerId(), "service response", booking.getId());
+        BaseRoleEntity provider = notificationHelper.getProvider(booking, BookingType.SERVICE);
+
+        String content = String.format(
+                EmailMessages.SERVICE_STATUS_UPDATE,
+                organizer.getFirstName(),
+                booking.getId(),
+                booking.getStatus(),
+                provider.getFirstName()
+        );
+        notificationHelper.send(organizer.getEmail(), "Service Booking Update", content);
+
     }
 }
