+++ src/main/java/com/example/cdr/eventsmanagementsystem/Service/Service/ServicesService.java
@@ -8,6 +8,7 @@ import com.example.cdr.eventsmanagementsystem.Model.Booking.Booking;
 import com.example.cdr.eventsmanagementsystem.Model.Booking.BookingStatus;
 import com.example.cdr.eventsmanagementsystem.Model.Service.Availability;
 import com.example.cdr.eventsmanagementsystem.Model.User.ServiceProvider;
+import com.example.cdr.eventsmanagementsystem.NotificationEvent.BookingUpdates.ServiceBookingUpdate;
 import com.example.cdr.eventsmanagementsystem.Repository.BookingRepository;
 import com.example.cdr.eventsmanagementsystem.Repository.ServiceRepository;
 import com.example.cdr.eventsmanagementsystem.Model.Service.Services;
@@ -15,6 +16,7 @@ import com.example.cdr.eventsmanagementsystem.Service.Auth.UserSyncService;
 import com.example.cdr.eventsmanagementsystem.Util.AuthUtil;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.AccessDeniedException;
@@ -27,14 +29,14 @@ import org.springframework.transaction.annotation.Transactional;
  */
 @RequiredArgsConstructor
 @Service
-public class ServicesService implements ServicesServiceInterface{
+public class ServicesService {
     private final ServiceRepository serviceRepository;
     private final BookingRepository bookingRepository;
     private final UserSyncService userSyncService;
     private final ServiceMapper serviceMapper;
     private final BookingMapper bookingMapper;
+    private final ApplicationEventPublisher eventPublisher;
 
-    @Override
     @Transactional
     public ServicesDTO addService(ServicesDTO dto) {
         Services newService = serviceMapper.toService(dto);
@@ -46,7 +48,6 @@ public class ServicesService implements ServicesServiceInterface{
         return serviceMapper.toServiceDTO(savedService);
     }
 
-    @Override
     @Transactional
     public ServicesDTO updateAvailability(Long serviceId) {
         String keycloakId = AuthUtil.getCurrentUserId();
@@ -54,7 +55,7 @@ public class ServicesService implements ServicesServiceInterface{
                 .orElseThrow(() -> new IllegalArgumentException("Service not found"));
 
         if (!service.getServiceProvider().getKeycloakId().equals(keycloakId)) {
-            throw new AccessDeniedException("You are not allowed to update availability of this service");
+            throw new AccessDeniedException("You are not allowed to update this service");
         }
         service.setAvailability(service.getAvailability() == Availability.AVAILABLE
                 ? Availability.UNAVAILABLE
@@ -64,14 +65,12 @@ public class ServicesService implements ServicesServiceInterface{
     }
 
     /// Refactor to its booking service
-    @Override
     public Page<BookingDetailsResponse> getBookingsForServiceProvider(Pageable pageable) {
         String serviceProviderId = AuthUtil.getCurrentUserId();
         Page<Booking> bookings = bookingRepository.findByService_ServiceProvider_Id(serviceProviderId,pageable);
         return bookings.map(bookingMapper::toBookingDetailsResponse);
     }
     /// Refactor to its booking service
-    @Override
     @Transactional
     public Booking respondToBookingRequests (Long bookingId, BookingStatus status) {
         String serviceProviderId = AuthUtil.getCurrentUserId();
@@ -82,20 +81,8 @@ public class ServicesService implements ServicesServiceInterface{
             throw new AccessDeniedException("You are not allowed to respond to this service");
         }
         booking.setStatus(status);
-        return bookingRepository.save(booking);
-    }
-    /// Refactor to its booking service
-    @Override
-    @Transactional
-    public void cancelBooking(Long bookingId){
-        String serviceProviderId = AuthUtil.getCurrentUserId();
-        Booking booking = bookingRepository.findById(bookingId)
-                .orElseThrow(() -> new EntityNotFoundException("Booking not found"));
-
-        if(!booking.getService().getServiceProvider().getKeycloakId().equals(serviceProviderId)) {
-            throw new AccessDeniedException("You are not allowed to cancel this service");
-        }
-        booking.setStatus(BookingStatus.CANCELLED);
-        bookingRepository.save(booking);
+        Booking savedBooking = bookingRepository.save(booking);
+        eventPublisher.publishEvent(new ServiceBookingUpdate(savedBooking));
+        return savedBooking;
     }
 }
